+++ lib/src/models/phone_number.dart
@@ -4,16 +4,6 @@ import 'package:phone_numbers_parser/src/parsers/validator.dart';
 import '../parsers/phone_number_parser.dart';
 import 'country.dart';
 
-/// {@template nationalNumber}
-/// The [nationalNumber] is the national part of a phone number.
-/// For example for the international phone number +33 93 987 6218
-/// the nationalNumber would be 93 987 6218, or alternatively
-/// with the prefix 0 93 987 6218.
-///
-/// Note that [PhoneNumber.nsn] is does not necessarily equal nationalNumber,
-/// as some manipulation is done on it.
-/// {@endtemplate}
-
 /// Represents a phone number.
 ///
 /// To know if a phone number is valid you can check the [valid] property.
@@ -21,15 +11,24 @@ import 'country.dart';
 /// The [international] property gives you a format that can be used to call
 /// a phone number.
 class PhoneNumber {
-  late final String nsn;
-  late final Country country;
+  /// The country this phone number originates from
+  final Country country;
+
+  /// The national number given as input, marked as private atm
+  /// because it could lead to confusion
+  final String _national;
+
+  /// The national number transformed for international use
+  final String nsn;
+
+  /// Whether this phone number is valid for its country
   late final bool valid;
 
   String get dialCode => country.dialCode;
   String get isoCode => country.isoCode;
   String get international => '+' + dialCode + nsn;
 
-  PhoneNumber._(this.country, this.nsn) {
+  PhoneNumber._(this.country, this.nsn, this._national) {
     valid = Validator.isValidNationalNumber(nsn, country.phone);
   }
 
@@ -59,7 +58,16 @@ class PhoneNumber {
 
   /// Creates a [PhoneNumber] with an isoCode or a nationalNumber
   ///
-  /// {@macro nationalNumber}
+  /// {@template nationalNumber}
+  ///
+  /// The [nationalNumber] is the national part of a phone number.
+  /// For example for the international phone number +33 93 987 6218
+  /// the nationalNumber would be 93 987 6218, or alternatively
+  /// with the prefix 0 93 987 6218.
+  ///
+  /// Note that [PhoneNumber.nsn] does not necessarily equal nationalNumber,
+  /// as some manipulation is done on it to make it valid internationally.
+  /// {@endtemplate}
   ///
   /// If the [isoCode] is invalid, throws a [PhoneNumberException]
   static PhoneNumber fromIsoCode(String isoCode, String nationalNumber) {
@@ -83,8 +91,13 @@ class PhoneNumber {
 
   /// This method does not transform the national number and should
   /// be used mainly by the library
-  static PhoneNumber fromCountry(Country country, String nationalNumber) {
-    return PhoneNumber._(country, nationalNumber);
+  static PhoneNumber fromCountry(
+      Country country, String nsn, String nationalNumber) {
+    return PhoneNumber._(country, nsn, nationalNumber);
+  }
+
+  PhoneNumber copyWithIsoCode(String isoCode) {
+    return PhoneNumberParser.parseWithIsoCode(_national, dialCode);
   }
 
   @override