+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - added toString to PhoneNumber
 - added leading digits getter to country
 - added displayDialCode on country which is a mix of dial code and leading digits
+- added copyWithIsoCode as instance method on phone number to make it
+  easy to switch between countries in a phone number input
 
 ## 0.0.2